started moving to new syscall memory model

+++ userland/aos_syscalls.h
@@ -16,7 +16,7 @@ u64 AOS_surface_commit(u64 surface_slot);
 u64 AOS_surface_acquire(u64 surface_slot, AOS_Framebuffer* fb, u64 page_count);
  
 u64 AOS_thread_awake_after_time(u64 duration);
-u64 AOS_thread_awake_on_surface(u64* surface_slots, u64 count);
+u64 AOS_thread_awake_on_surface(u16* surface_slots, u64 count);
 
 typedef struct
 {
